+++ CHANGELOG.md
@@ -81,6 +81,7 @@ or port 0, but it was impossible to remove it.
 - The widget that displayed IRC network port when connecting to IRC network
 had a minimum value set to 6667. If IRC network used a port lower than 6667,
 it would not display properly.
+- Remote Console connect dialog had no title.
 - Linux: The name of the desktop file is now a reverse Doomseeker domain.
 - Zandronum: fix problem with widgets for a different Zandronum version being
 visible when first opening the "Create Game" box.